backend/requirements.txt
astroid==3.0.2
blinker==1.7.0
click==8.1.7
dill==0.3.7
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.0.3
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
msgpack==1.0.7
neovim==0.3.1
numpy==1.26.2
Pillow==10.1.0
platformdirs==4.1.0
pydicom==2.4.4
pylint==3.0.3
pynvim==0.5.0
python-magic==0.4.27
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
Werkzeug==3.0.1
